+++ noodles-gff/src/io/writer/line/directive.rs
@@ -3,7 +3,10 @@ mod value;
 use std::io::{self, Write};
 
 use self::value::write_value;
-use crate::{directive_buf::key, DirectiveBuf};
+use crate::{
+    directive_buf::{key, Value},
+    DirectiveBuf,
+};
 
 pub(crate) fn write_directive<W>(writer: &mut W, directive: &DirectiveBuf) -> io::Result<()>
 where
@@ -11,29 +14,35 @@ where
 {
     write_prefix(writer)?;
 
-    match directive {
-        DirectiveBuf::GffVersion(version) => {
+    match (directive.key(), directive.value()) {
+        (key::GFF_VERSION, Some(Value::GffVersion(version))) => {
             write_key(writer, key::GFF_VERSION)?;
             write_separator(writer)?;
             value::write_gff_version(writer, version)?;
         }
-        DirectiveBuf::SequenceRegion(sequence_region) => {
+        (key::SEQUENCE_REGION, Some(Value::SequenceRegion(sequence_region))) => {
             write_key(writer, key::SEQUENCE_REGION)?;
             write_separator(writer)?;
             value::write_sequence_region(writer, sequence_region)?;
         }
-        DirectiveBuf::GenomeBuild(genome_build) => {
+        (key::GENOME_BUILD, Some(Value::GenomeBuild(genome_build))) => {
             write_key(writer, key::GENOME_BUILD)?;
             write_separator(writer)?;
             value::write_genome_build(writer, genome_build)?;
         }
-        DirectiveBuf::Other(key, value) => {
+        (key, Some(Value::Other(value))) => {
             write_key(writer, key)?;
-
-            if let Some(v) = value {
-                write_separator(writer)?;
-                write_value(writer, v)?;
-            }
+            write_separator(writer)?;
+            write_value(writer, value)?;
+        }
+        (key, None) => {
+            write_key(writer, key)?;
+        }
+        _ => {
+            return Err(io::Error::new(
+                io::ErrorKind::InvalidInput,
+                "invalid directive",
+            ))
         }
     }